comment 도메인 삭제 admin으로 변경
