+++ src/main/resources/templates/admin/inquiryListDetail.html
@@ -77,7 +77,7 @@
         <div class="comment_admin card my-4" th:if="${comments.isEmpty()}">
             <h5 class="card-header">답변 쓰기</h5>
             <div class="card-body">
-                <form th:action="@{/{id}/comment(id=${id})}" th:method="post">
+                <form th:action="@{/admin/{id}/comment(id=${id})}" th:method="post">
                     <div class="form-group" style="margin-bottom: 10px;">
                         <input type="hidden" name="idx" th:value="*{idx}"/>
                         <textarea name="content" class="form-control" rows="3"></textarea>
@@ -102,7 +102,7 @@
                                 <button class="btn btn-sm btn-outline-warning" th:attr="data-comment_id=${comment.id}, data-board_id=${board.id}"
                                         onclick="editComment(this)">수정
                                 </button>
-                                <a th:href="@{/board/{id}/comment/{commentId}/remove(id=${id}, commentId=${comment.id})}"
+                                <a th:href="@{/admin/{id}/comment/{commentId}/remove(id=${id}, commentId=${comment.id})}"
                                    class="btn btn-sm btn-outline-danger"
                                    onclick="return confirm('답변을 삭제하시겠습니까?')">삭제</a>
                             </div>
